Refactor webpack config generation
