+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/getNodeEnv.ts
@@ -1,0 +1,5 @@
+export function getNodeEnv() {
+  // Disallow non dev/prod environments, like "test" inside Jest, because
+  // they are not supported by webpack
+  return process.env.NODE_ENV === 'production' ? 'production' : 'development';
+}
