+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/getWebpackConfigResolve.ts
@@ -1,0 +1,72 @@
+import path from 'path';
+import resolveFrom from 'resolve-from';
+import webpack from 'webpack';
+import { CosmosConfig } from '../../../config';
+
+export function getWebpackConfigResolve(
+  cosmosConfig: CosmosConfig,
+  webpackConfig: webpack.Configuration
+): webpack.ResolveOptions {
+  return resolveLocalReactDeps(cosmosConfig, webpackConfig);
+}
+
+function resolveLocalReactDeps(
+  cosmosConfig: CosmosConfig,
+  webpackConfig: webpack.Configuration
+): webpack.ResolveOptions {
+  const { rootDir } = cosmosConfig;
+
+  const { resolve = {} } = webpackConfig;
+  let alias = resolve.alias || {};
+
+  // Preserve existing React aliases (eg. when using Preact)
+  let reactAlias = hasAlias(alias, 'react');
+  let reactDomAlias = hasAlias(alias, 'react-dom');
+
+  if (reactAlias && reactDomAlias) {
+    console.log('[Cosmos] React and React DOM aliases found in webpack config');
+    return resolve;
+  }
+
+  if (reactAlias) {
+    console.log('[Cosmos] React alias found in webpack config');
+  } else {
+    const reactPath = resolveFrom.silent(rootDir, 'react');
+    if (!reactPath)
+      throw new Error(`[Cosmos] Local dependency not found: react`);
+    alias = addAlias(alias, 'react', path.dirname(reactPath));
+  }
+
+  if (reactDomAlias) {
+    console.log('[Cosmos] React DOM alias found in webpack config');
+  } else {
+    const reactDomPath = resolveFrom.silent(rootDir, 'react-dom');
+    if (!reactDomPath)
+      throw new Error(`[Cosmos] Local dependency not found: react-dom`);
+    alias = addAlias(alias, 'react-dom', path.dirname(reactDomPath));
+  }
+
+  return { ...resolve, alias };
+}
+
+function hasAlias(alias: webpack.ResolveOptions['alias'], name: string) {
+  if (!alias) return false;
+
+  const exactName = `${name}$`;
+  if (Array.isArray(alias)) {
+    return alias.some(a => a.name === name || a.name === exactName);
+  } else {
+    const keys = Object.keys(alias);
+    return keys.includes(name) || keys.includes(exactName);
+  }
+}
+
+function addAlias(
+  alias: webpack.ResolveOptions['alias'],
+  name: string,
+  value: string | false | string[]
+) {
+  return Array.isArray(alias)
+    ? [...alias, { name, alias: value }]
+    : { ...alias, [name]: value };
+}
