+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/plugins.ts
@@ -1,0 +1,39 @@
+import webpack from 'webpack';
+import { CosmosConfig } from '../../../config';
+import { getNodeEnv } from './getNodeEnv';
+
+export function getGlobalsPlugin(
+  { publicUrl }: CosmosConfig,
+  userWebpack: typeof webpack,
+  devServerOn: boolean
+) {
+  const cleanPublicUrl = removeTrailingSlash(publicUrl);
+  return new userWebpack.DefinePlugin({
+    // "if (__DEV__)" blocks get stripped when compiling a static export build
+    __DEV__: JSON.stringify(devServerOn),
+    'process.env': {
+      NODE_ENV: JSON.stringify(getNodeEnv()),
+      PUBLIC_URL: JSON.stringify(cleanPublicUrl),
+    },
+  });
+}
+
+export function hasPlugin(
+  plugins: void | webpack.WebpackPluginInstance[],
+  pluginName: string
+) {
+  return (
+    plugins && plugins.filter(p => isInstanceOfPlugin(p, pluginName)).length > 0
+  );
+}
+
+export function isInstanceOfPlugin(
+  plugin: webpack.WebpackPluginInstance,
+  constructorName: string
+) {
+  return plugin.constructor && plugin.constructor.name === constructorName;
+}
+
+function removeTrailingSlash(url: string) {
+  return url.replace(/\/$/, '');
+}
