+++ packages/react-cosmos/src/plugins/webpack/webpackConfig/resolveWebpackClientPath.ts
@@ -1,0 +1,3 @@
+export function resolveWebpackClientPath(relPath: string) {
+  return require.resolve(`../client/${relPath}`);
+}
